codegen: refactoring
