+++ src/tools/kdb/CMakeLists.txt
@@ -2,7 +2,8 @@ include (LibAddMacros)
 
 file (GLOB HDR_FILES
 	   *.hpp
-	   gen/*.hpp)
+	   gen/*.hpp
+	   gen/elektra/*.hpp)
 add_headers (HDR_FILES)
 add_cppheaders (HDR_FILES)
 add_toolheaders (HDR_FILES)
@@ -10,7 +11,8 @@ include_directories (${CMAKE_CURRENT_SOURCE_DIR})
 
 file (GLOB SRC_FILES
 	   *.cpp
-	   gen/*.cpp)
+	   gen/*.cpp
+	   gen/elektra/*.cpp)
 set (SOURCES ${SRC_FILES} ${HDR_FILES})
 
 add_subdirectory (gen/templates)
