+++ src/tools/kdb/gen/elektra/elektragen.hpp
@@ -9,7 +9,7 @@
 #ifndef ELEKTRA_ELEKTRAGEN_HPP
 #define ELEKTRA_ELEKTRAGEN_HPP
 
-#include "template.hpp"
+#include "../template.hpp"
 
 class ElektraGenTemplate : public GenTemplate
 {
